sbt: add a wrapJdk option
The current behaviour is to hardcode the jdk by default. This is
practical because it allows sbt to run as a standalone executable
but can cause surprising behaviour when a different jdk
version is in the scope (via nix-shell -p jdk* for instance), as the jdk
will not be changed automatically.

This allows a user to choose not to wrap the jdk, at the cost of having
to make sure a jdk is in the path by another mean.